# docs/api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "科研助手 sidecar",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/run": {
      "post": {
        "summary": "Run",
        "operationId": "run_api_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deai/scan": {
      "post": {
        "summary": "Deai Scan",
        "description": "去 AI 味 第一步: 启发式扫描, 标出 AI 味较重的句子(非流式, 不调 LLM)。\n\n任何异常都返回空结果(放行不阻塞), 与其它\"分析型\"接口一致。",
        "operationId": "deai_scan_api_deai_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeaiScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deai/rewrite": {
      "post": {
        "summary": "Deai Rewrite Ep",
        "description": "去 AI 味 第二步: 逐块流式改写被标记的散文块(可随时断流中断)。",
        "operationId": "deai_rewrite_ep_api_deai_rewrite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeaiRewriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/idea": {
      "post": {
        "summary": "Idea",
        "description": "医学/药学/生物 找选题: 检索 PubMed + 分析现状/空白/选题(带文献链接)。",
        "operationId": "idea_api_idea_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/idea/clarify": {
      "post": {
        "summary": "Idea Clarify",
        "description": "检索前澄清: 判断方向是否够具体, 不够则回最多 3 个澄清问题(非流式)。\n\n任何异常/mock 一律返回 ready=True 放行, 绝不阻塞用户检索。",
        "operationId": "idea_clarify_api_idea_clarify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/idea/refine": {
      "post": {
        "summary": "Idea Refine",
        "description": "澄清回答后, 给出 2-3 个优化的研究方向/关键词候选(非流式)。失败/mock 返回空 options 放行。",
        "operationId": "idea_refine_api_idea_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imrad": {
      "post": {
        "summary": "Imrad Ep",
        "description": "IMRaD 初稿装配: 把已有材料分段拼成 Intro/Methods/Results/Discussion(本地, 只据材料)。",
        "operationId": "imrad_ep_api_imrad_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/poster": {
      "post": {
        "summary": "Poster Ep",
        "description": "学术海报: 把论文提炼成海报要点(纯文本 LLM) + 确定性渲染自包含 HTML(可打印 PDF)。",
        "operationId": "poster_ep_api_poster_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/poster/review": {
      "post": {
        "summary": "Poster Review Ep",
        "description": "VLM 排版审阅: 看海报截图找排版问题, 返回评审意见 + 修订后的要点与重渲染 HTML(非流式)。",
        "operationId": "poster_review_ep_api_poster_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grant": {
      "post": {
        "summary": "Grant Ep",
        "description": "写中文标书: 凝练方案 → 大纲 → 分节撰写 → 评审自查(据选题报告与真实文献, 接在找选题之后)。\n\n若 inputs 含已确认的 scheme/sections, 则跳过凝练直接据此撰写(两段式)。",
        "operationId": "grant_ep_api_grant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grant/search": {
      "post": {
        "summary": "Grant Search Ep",
        "description": "写标书 Step 1.5: 单独跑一次文献检索 + 核心发现抽取 (供 picker 使用)。",
        "operationId": "grant_search_ep_api_grant_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grant/plan": {
      "post": {
        "summary": "Grant Plan Ep",
        "description": "两段式第一步: 产出可编辑的【方案骨架 + 大纲】交用户确认(非流式)。失败也回退不阻断。",
        "operationId": "grant_plan_ep_api_grant_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grant/style": {
      "post": {
        "summary": "Grant Style Ep",
        "description": "从上传的文风样例提炼一份『文风档案』(非流式)。失败/空样例返回空档案, 不阻断撰写。",
        "operationId": "grant_style_ep_api_grant_style_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantStyleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grant/review": {
      "post": {
        "summary": "Grant Review Ep",
        "description": "对当前申请书全文重新跑一遍评审组模拟评审(修订后回头看改进了没), 流式。",
        "operationId": "grant_review_ep_api_grant_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grant/revise": {
      "post": {
        "summary": "Grant Revise Ep",
        "description": "逐节重写: 仅按意见重写某一章节, 不重跑全篇(流式)。",
        "operationId": "grant_revise_ep_api_grant_revise_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rebuttal": {
      "post": {
        "summary": "Rebuttal Ep",
        "description": "回复审稿意见: 拆解意见 → 逐条生成 point-by-point 回复信(本地处理, 数据不出网)。",
        "operationId": "rebuttal_ep_api_rebuttal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeaiRewriteRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "blocks": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Blocks",
            "default": []
          },
          "style": {
            "type": "string",
            "title": "Style",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DeaiRewriteRequest"
      },
      "DeaiScanRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DeaiScanRequest"
      },
...
      "GrantStyleRequest": {
        "properties": {
          "sample": {
            "type": "string",
            "title": "Sample"
          }
        },
        "type": "object",
        "required": [
          "sample"
        ],
        "title": "GrantStyleRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunRequest": {
        "properties": {
          "module": {
            "type": "string",
            "title": "Module"
          },
          "inputs": {
            "type": "object",
            "title": "Inputs"
          }
        },
        "type": "object",
        "required": [
          "module",
          "inputs"
        ],
        "title": "RunRequest"
      },
...
    }
  }
}